Bugfixing and added recommended thread and block settings and a device query option

+++ cpu_code.h
@@ -1,8 +1,8 @@
 #pragma once
 
 void tryKeyRandom(int i, char *nonce, char*veribuf);
-void measureCPUPerformance(unsigned long long nrOfThreads,
-						   unsigned long long *nrKeysCalculatedResult,
-						   unsigned long long *nrOfSecondsInTotalMeasured,
+void measureCPUPerformance(uint64_t nrOfThreads,
+						   uint64_t *nrKeysCalculatedResult,
+						   uint64_t *nrOfSecondsInTotalMeasured,
 						   bool* shutdownRequested,
-						   unsigned long long nrSecondsToMeasure = 30);
+						   uint64_t nrSecondsToMeasure = 30);
